+++ src/main/java/com/zl/base/MyMapper.java
@@ -1,4 +1,4 @@
-package com.hdy.base;
+package com.zl.base;
 
 import tk.mybatis.mapper.common.Mapper;
 import tk.mybatis.mapper.common.MySqlMapper;
