+++ src/main/java/com/zl/mapper/datasource2/TCountryMapper2.java
@@ -1,6 +1,8 @@
-package com.hdy.mapper.external;
+package com.zl.mapper.datasource2;
 
-import com.hdy.model.external.TCountry;
+
+import com.zl.model.datasource2.TCountry;
+
 import tk.mybatis.mapper.common.Mapper;
 
 public interface TCountryMapper2 extends Mapper<TCountry> {
