+++ src/main/resources/config.properties
@@ -25,17 +25,19 @@
 # projest package
 #project=D:/workspace-sts/spring4-mybatis3/src/main
 #project=/Volumes/IntelSSD/Prj/workspace-sts/springmvc-mybatis-rest-pro/src/main
-prj.groupId1=com.hdy
-prj.groupId2=com/hdy
+prj.groupId1=com.zl
+prj.groupId2=com/zl
 
 
 # \u6570\u636E\u5E93\u914D\u7F6E
+# datasource1
 jdbc.driverClass = com.mysql.jdbc.Driver
 jdbc.url = jdbc:mysql://127.0.0.1:3306/zea
 jdbc.user = root
 jdbc.password =zealous
 
 # \u6570\u636E\u5E93\u914D\u7F6E
+# datasource2
 jdbc.driverClass2 = com.mysql.jdbc.Driver
 jdbc.url2 = jdbc:mysql://192.168.0.248:3306/zea
 jdbc.user2 = ogg
