+++ src/main/resources/generator/database.sql
@@ -1,5 +1,5 @@
 
-mvn archetype:generate -DarchetypeArtifactId=maven-archetype-webapp -DinteractiveMode=false -DgroupId=com.hdy -DartifactId=springmvc-mybatis-rest-pro
+mvn archetype:generate -DarchetypeArtifactId=maven-archetype-webapp -DinteractiveMode=false -DgroupId=com.zl -DartifactId=springmvc-mybatis-rest-pro
 
 USE `zea`;
