+++ src/main/resources/mapper/datasource2/TCountryMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="com.hdy.mapper.external.TCountryMapper2" >
-  <resultMap id="BaseResultMap" type="com.hdy.model.external.TCountry" >
+<mapper namespace="com.zl.mapper.datasource2.TCountryMapper2" >
+  <resultMap id="BaseResultMap" type="com.zl.model.datasource2.TCountry" >
     <!--
       WARNING - @mbggenerated
     -->
